+++ packages/decap-cms-backend-git-gateway/src/implementation.ts
@@ -1,6 +1,8 @@
 import GoTrue from 'gotrue-js';
 import jwtDecode from 'jwt-decode';
-import { get, pick, intersection } from 'lodash';
+import get from 'lodash/get';
+import pick from 'lodash/pick';
+import intersection from 'lodash/intersection';
 import ini from 'ini';
 import {
   APIError,
@@ -362,7 +364,11 @@ export default class GitGateway implements Implementation {
       if (!(await this.api!.hasWriteAccess())) {
         throw new Error("You don't have sufficient permissions to access Decap CMS");
       }
-      return { name: userData.name, login: userData.email } as User;
+      return {
+        name: userData.name,
+        login: userData.email,
+        avatar_url: userData.avatar_url,
+      } as unknown as User;
     });
   }
   async restoreUser() {