+++ packages/decap-cms-backend-proxy/src/AuthenticationPage.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import styled from '@emotion/styled';
-import { Icon, buttons, shadows, GoBackButton } from 'decap-cms-ui-default';
+import { Icon, buttons, shadows, GoBackButton, renderPageLogo } from 'decap-cms-ui-default';
 
 const StyledAuthenticationPage = styled.section`
   display: flex;
@@ -9,11 +9,6 @@ const StyledAuthenticationPage = styled.section`
   align-items: center;
   justify-content: center;
   height: 100vh;
-`;
-
-const PageLogoIcon = styled(Icon)`
-  color: #c4c6d2;
-  margin-top: -300px;
 `;
 
 const LoginButton = styled.button`
@@ -41,6 +36,16 @@ export default class AuthenticationPage extends React.Component {
     t: PropTypes.func.isRequired,
   };
 
+  componentDidMount() {
+    // Manually validate PropTypes - React 19 breaking change
+    PropTypes.checkPropTypes(
+      AuthenticationPage.propTypes,
+      this.props,
+      'prop',
+      'AuthenticationPage',
+    );
+  }
+
   handleLogin = e => {
     e.preventDefault();
     this.props.onLogin(this.state);
@@ -51,7 +56,7 @@ export default class AuthenticationPage extends React.Component {
 
     return (
       <StyledAuthenticationPage>
-        <PageLogoIcon size="300px" type="decap-cms" />
+        {renderPageLogo(config.logo_url)}
         <LoginButton disabled={inProgress} onClick={this.handleLogin}>
           {inProgress ? t('auth.loggingIn') : t('auth.login')}
         </LoginButton>