+++ packages/decap-cms-lib-widgets/src/validations.ts
@@ -1,4 +1,4 @@
-import { isNumber } from 'lodash';
+import isNumber from 'lodash/isNumber';
 
 import type { List } from 'immutable';