+++ packages/decap-cms-widget-markdown/src/MarkdownControl/index.js
@@ -54,6 +54,11 @@ export default class MarkdownControl extends React.Component {
     };
   }
 
+  componentDidMount() {
+    // Manually validate PropTypes - React 19 breaking change
+    PropTypes.checkPropTypes(MarkdownControl.propTypes, this.props, 'prop', 'MarkdownControl');
+  }
+
   handleMode = mode => {
     this.setState({ mode, pendingFocus: true });
     localStorage.setItem(MODE_STORAGE_KEY, mode);
@@ -66,6 +71,10 @@ export default class MarkdownControl extends React.Component {
   };
 
   getAllowedModes = () => this.props.field.get('modes', List(['rich_text', 'raw'])).toArray();
+
+  focus() {
+    this.setState({ pendingFocus: true });
+  }
 
   render() {
     const {