+++ packages/decap-cms-widget-string/src/StringControl.js
@@ -29,6 +29,11 @@ export default class StringControl extends React.Component {
   // The input element ref
   _el = null;
 
+  componentDidMount() {
+    // Manually validate PropTypes - React 19 breaking change
+    PropTypes.checkPropTypes(StringControl.propTypes, this.props, 'prop', 'StringControl');
+  }
+
   // NOTE: This prevents the cursor from jumping to the end of the text for
   // nested inputs. In other words, this is not an issue on top-level text
   // fields such as the `title` of a collection post. However, it becomes an