+++ packages/decap-cms-locales/CHANGELOG.md
@@ -3,6 +3,26 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [3.4.0](https://github.com/decaporg/decap-cms/compare/decap-cms-locales@3.3.0...decap-cms-locales@3.4.0) (2025-06-26)
+
+### Features
+
+- update locale (it) ([#7508](https://github.com/decaporg/decap-cms/issues/7508)) ([30c4fd2](https://github.com/decaporg/decap-cms/commit/30c4fd293aba9672eb4be5a6a4517340e01ca2dc))
+
+# [3.3.0](https://github.com/decaporg/decap-cms/compare/decap-cms-locales@3.2.0...decap-cms-locales@3.3.0) (2025-01-15)
+
+### Features
+
+- add mk (Macedonian) locale ([#7346](https://github.com/decaporg/decap-cms/issues/7346)) ([b41e43a](https://github.com/decaporg/decap-cms/commit/b41e43a73218f0950ea9c96f21dfe0a31aa173b3))
+- **locale/cs:** improve the Czech translations, add new strings from `en` ([#7326](https://github.com/decaporg/decap-cms/issues/7326)) ([76f4b74](https://github.com/decaporg/decap-cms/commit/76f4b7411684943c515a77932fcac743adce8733))
+
+# [3.2.0](https://github.com/decaporg/decap-cms/compare/decap-cms-locales@3.1.4...decap-cms-locales@3.2.0) (2024-08-07)
+
+### Bug Fixes
+
+- **i18n:** improve Thai locale ([#7248](https://github.com/decaporg/decap-cms/issues/7248)) ([65b9349](https://github.com/decaporg/decap-cms/commit/65b9349c105f2b98dd0e4d4f19ceb0d7a6ee293a))
+- update ukrainian lang ([#7252](https://github.com/decaporg/decap-cms/issues/7252)) ([9881de0](https://github.com/decaporg/decap-cms/commit/9881de0587925840cd15f3af0838284a0ca2b0ab))
+
 ## [3.1.4](https://github.com/decaporg/decap-cms/compare/decap-cms-locales@3.1.3...decap-cms-locales@3.1.4) (2024-03-21)
 
 **Note:** Version bump only for package decap-cms-locales